Implemented OSX support including a package installer and uninstaller.
Added a file that only contains the URL to the config interface instead of having to read through the logs to find it. Started writing the PID of any process that holds an exclusive repository lock.

Fixed a bug with file locking where exclusive and shared were mixed up. Also fixed a bug where pending active paths were not properly flushed on shutdown. Fixed an issue restoring without writing files.

Finish a backup by reporting back statistics of the contents of the backup to the UI.

Added an option to schedule repository optimization. Defaults to once a month for new installs. Also added option to not have a schedule on backup sets and only run when invoked manually.