Added an option to ignore any log entries to delete data from repository when rebuilding repository from logs. This can allow you to recover from certain kinds of failures. However, use with care since if the blocks referenced were also deleted from the destination the metadata would be addressing data that no longer exists (Can only be verified by doing a restore).
Also added some additional guard rails against invalid super blocks and also fixed a security issue with the newly introduced direct download feature. Previous when using this the repository password was passed as a query parameter which means it would also often be stored in the history of your browser. Now the direct download uses a POST and this issue is resolved.