Added a very small but useful feature to do a quick search for contents of the repository through the command line using the new `search` command. Also added the ability to list full path names when doing an `ls` operation.
Finally added the option to set a maximum retention for a destination. If set the application will read and rewrite any blocks that have been untouched after the maximum retention duration have been reached.

Fixed a bug relating to block validation when multiple destinations for a single set were used.