An issue relating to directory contents sometimes not being stored at the end of a backup in case its contents changed during the backup scan was fixed. The issue was not critical since it would always be corrected on the next run and the data would still show up if you included deleted contents.
Another issue was discovered where if the upload of a log file failed it would cause the application to be in an invalid state where no more logs would be generated until the application was restarted. Although critical it was a very unlikely case (But it did eventually end up happening).

Both of these are fixed in the new version.