Fixed a critical bug with UNICODE when processing logs.
On top of this another minor bug was fixed where sometimes log files were not properly deleted after upload. Another bug was fixed for handling paths that are invalid properly without just crashing on them.

Additionally some optimizations to when trimming occurs and how thorough it is. A new advanced option has been introduced to control how often the repository should be processed to find unused blocks in the destination. This option can be useful if you have a very large repository where the block validation could start taking a significant time.