Fixes for two extremely rare but serious issues.
First there was an issue with modifying a list without proper locking causing the application to hang in rare cases.

Secondly there was an even rare but possible case where two different blocks with different contents could calculate to the same block hash when using different block assignment types. I have never seen this happen in the wild, but it was possible and would lead to minor data corruption.